==> paper-analysis/191010-PredictStabilityInMZT/01_predict_decay_during_MZT.ipynb <==
import pandas as pd
import numpy as np
from joblib import load

from optimalcodon.projects.rnastability.dataprocessing import get_data, general_preprocesing_pipeline
from optimalcodon.projects.rnastability.predictdecay import predict_sequence

(train_x, train_y), (test_x, test_y) = get_data(
    "../data/191004-TrainAndTestSets/")

# TODO: replace this line with the new model (this model is overfitting)
model = load("../191005-EvaluateModelLearningCurve/results-data/final_model.joblib")


# load the data for fish
fish_data = pd.read_csv("../../data/19-01-17-Get-ORFS-UTRS-codon-composition/sequence-data/fish_seq_data_cds_3utr.csv")
fish_data['specie'] = "fish"
fish_data['cell_type'] = "embryo mzt"
fish_data['datatype'] = "aamanitin polya"
fish_data['utrlenlog'] = np.log(fish_data['3utr'].str.len() + 1)
fish_data['cdslenlog'] = np.log(fish_data['coding'].str.len() + 1)
# rename the undex

fish_data = (
    fish_data.
    rename(columns={'ensembl_gene_id': 'gene_id'})
    .set_index('gene_id')
    .drop(['3utr'], axis=1)
    .loc[:, train_x.columns.values]
)


fish_data.sample(5)

# load xen data
xen_data = pd.read_csv("../../data/19-06-12-XenopusData/xenopus_seq_data_cds_3utr.csv")
xen_data['specie'] = "xenopus"
xen_data['cell_type'] = "embryo mzt"
xen_data['datatype'] = "aamanitin polya"
xen_data['utrlenlog'] = np.log(xen_data['3utr'].str.len() + 1)
xen_data['cdslenlog'] = np.log(xen_data['coding'].str.len() + 1)

xen_data = (
    xen_data.
    rename(columns={'ensembl_gene_id': 'gene_id'})
    .set_index('gene_id')
    .drop(['3utr'], axis=1)
    .loc[:, train_x.columns.values]
)
xen_data.sample(5)

fish_and_xen_data = pd.concat([xen_data, fish_data])
fish_and_xen_data.sample(10)

fish_and_xen_data['predicted_stability'] = model.predict(fish_and_xen_data)

fish_and_xen_data.to_csv("results-data/predicted_stabilit_fish_and_xenopus_all_genes.csv")
